Don’t allow deleting of static schedules

+++ apps/webapp/app/v3/services/deleteTaskSchedule.server.ts
@@ -27,9 +27,24 @@ export class DeleteTaskScheduleService extends BaseService {
     }
 
     try {
+      const schedule = await this._prisma.taskSchedule.findFirst({
+        where: {
+          friendlyId,
+        },
+      });
+
+      if (!schedule) {
+        throw new Error("Schedule not found");
+      }
+
+      if (schedule.type === "STATIC") {
+        throw new Error("Cannot delete static schedules");
+      }
+
       await this._prisma.taskSchedule.delete({
         where: {
           friendlyId,
+          type: "DYNAMIC",
         },
       });
     } catch (e) {
